# docs/openapi-v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Conectadois API",
    "version": "0.1.0-draft",
    "summary": "Contrato versionado do núcleo do MVP do Conectadois",
    "description": "Contrato-alvo da API v1. As extensões x-implementation-status e x-legacy-path distinguem o comportamento atual do servidor JSON das operações planejadas. A versão permanece draft até a decisão de mecânicas da tarefa 112 e a validação de segurança."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Mesmo host do frontend"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/couples/join": {
      "post": {
        "tags": [
          "Couples"
        ],
        "summary": "Aceitar convite e ocupar a segunda vaga",
        "operationId": "joinCouple",
        "x-implementation-status": "partial",
        "x-legacy-path": "POST /api/couples/join",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinCoupleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Casal ativado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoupleResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "410": {
            "$ref": "#/components/responses/Gone"
          }
        }
      }
    },
...
    "/love-notes": {
      "get": {
        "tags": ["Couples"],
        "summary": "Listar recados do casal",
        "operationId": "listLoveNotes",
        "x-implementation-status": "partial",
        "x-legacy-path": "GET /api/love-notes",
        "responses": {
          "200": {
            "description": "Recados visíveis somente aos integrantes vigentes do casal",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["notes"],
                  "properties": {
                    "notes": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/LoveNote" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      },
      "post": {
        "tags": ["Couples"],
        "summary": "Enviar recado ao parceiro",
        "operationId": "createLoveNote",
        "x-implementation-status": "partial",
        "x-legacy-path": "POST /api/love-notes",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoveNoteRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recado enviado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["note"],
                  "properties": {
                    "note": { "$ref": "#/components/schemas/LoveNote" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/game-sessions": {
      "post": {
        "tags": [
          "Games"
        ],
        "summary": "Criar partida bilateral com conteúdo congelado",
        "operationId": "createGameSession",
        "x-implementation-status": "partial",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Partida e rodadas criadas atomicamente",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSessionResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/game-sessions/{sessionId}": {
      "get": {
        "tags": [
          "Games"
        ],
        "summary": "Retomar partida autorizada",
        "operationId": "getGameSession",
        "x-implementation-status": "partial",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Partida com rodadas visíveis para a conta atual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameSessionResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Token opaco de sessão individual; nunca identifica o casal por parâmetro enviado pelo cliente."
      }
    },
...
    "parameters": {
      "SessionId": {
        "name": "sessionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "UUID único por intenção; mesma chave e payload repetem a resposta original por 24 horas.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "$ref": "#/components/responses/Problem400"
      },
      "Unauthorized": {
        "$ref": "#/components/responses/Problem401"
      },
      "Forbidden": {
        "$ref": "#/components/responses/Problem403"
      },
      "NotFound": {
        "$ref": "#/components/responses/Problem404"
      },
      "Conflict": {
        "$ref": "#/components/responses/Problem409"
      },
      "Gone": {
        "$ref": "#/components/responses/Problem410"
      },
...
      "ValidationError": {
        "$ref": "#/components/responses/Problem422"
      },
...
      "Problem400": {
        "description": "Requisição malformada",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem401": {
        "description": "Sessão ausente, inválida, expirada ou revogada",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Problem404": {
        "description": "Recurso não encontrado ou ocultado por segurança",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem409": {
        "description": "Estado atual impede a operação",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Problem410": {
        "description": "Token ou convite expirado, consumido ou revogado",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Problem422": {
        "description": "Campos semanticamente inválidos",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "JoinCoupleRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "invitationCode"
        ],
        "properties": {
          "invitationCode": {
            "type": "string",
            "pattern": "^[A-Z0-9-]{8,32}$"
          }
        }
      },
...
      "Couple": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "status",
          "anniversary",
          "timezone",
          "partner",
          "createdAt",
          "pairedAt",
          "invitation"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "waiting",
              "active",
              "closed"
            ]
          },
          "anniversary": {
            "oneOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ]
          },
          "timezone": {
            "type": "string"
          },
          "partner": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Partner"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "pairedAt": {
            "oneOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "invitation": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Invitation"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "CoupleResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Couple"
          }
        }
      },
...
      "Depth": {
        "type": "string",
        "enum": [
          "light",
          "medium",
          "deep"
        ]
      },
...
      "CreateGameSessionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mode",
          "roundCount"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "questions",
              "challenges",
              "mixed"
            ]
          },
          "roundCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          },
          "themes": {
            "type": "array",
            "maxItems": 20,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "maxLength": 50
            }
          },
          "depths": {
            "type": "array",
            "maxItems": 3,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/Depth"
            }
          }
        }
      },
...
      "LoveNoteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["text"],
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 280
          }
        }
      },
      "LoveNote": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "text",
          "senderName",
          "recipientName",
          "createdAt",
          "mine"
        ],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "text": { "type": "string", "minLength": 1, "maxLength": 280 },
          "senderName": { "type": "string" },
          "recipientName": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "mine": { "type": "boolean" }
        }
      },
      "GameSessionResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GameSession"
          }
        }
      },
...
    }
  }
}